requirements.txt
streamlit
torch
torchvision
transformers
faiss-cpu
pillow
numpy